.task.sets.subtask can now be used to pass around tasks as callbacks.
subtask is now a dict subclass, meaning it can safely be serialized
using JSON and other serializers that doesn't support complex objects.

Also if the first argument to subtask() is another subtask or a dict,
it will ignore the rest of the argument and use the values from the dict
instead, this means the task handling the callback can use subtask() like a
type to convert a subtask that has been serialized into a dict back into a
subtask::

    @task()
    def add(x, y, callback=None):
        result = x + y
        if callback:
            subtask(callback).apply_async(result)

subtask.apply_async now takes varargs that will be prepended to the stored
arguments, this is added for easy callback invocation. The new signature
is subtask.apply_async(*argmerge, **execopts)::

if callback is::

    callback = subtask("tasks.add", args=(10, ))

and callback is applied like::

    subtask(callback).apply_async(result)

The final arguments will be::

    (result, 10)

In addition to these changes the Task now has a new classmethod
to easily create subtasks:

    Task.subtask(args, kwargs, options)

which means we can launch the add task above with a callback like this::

    add.delay(16, 16, callback=add.subtask(args=(10, ))

this will result in (16 + 16) + 10
(allthough that is not the result stored by the first task id, rather
the second)