TaskSetResult.join + AsyncResult.wait improvements.
The implementation actually haven't changed since v0.4.0, and we can
simplify this a lot now.

An `interval` keyword argument has been added to both so the
polling interval can be specified.

A `propagate` keyword argument has been added to `result.wait()`,
errors will be included instead of raised if this is set to False.

Polling results when using the database backend is still probabably
not a good idea.