multiprocessing.Pool: Fixes race condition with WorkerLostError.
The process may have published a result before it was terminated,
but we have no reliable way to detect that this is the case.

So we have to wait for 10 seconds before marking the result with
WorkerLostError.  This gives the result handler a chance to retrieve the
result.  Downside is that it takes ~10 seconds for a job to be marked
with WorkerLostError, but I think it's a compromise we can live with,
at least it decreases the chance of having false negatives.

Closes #268.  Thanks to dialtone.