Verify that the pidfile is actually written correctly.  Closes #641
Hopefully this will crash the worker immediately if the system
is out of space to store the complete pidfile.

In addition, we now verify that existing pidfiles contain
a new line so that a partially written pidfile is detected as broken,
as before doing:

    echo -n "1" > celeryd.pid

would cause celeryd to think that an existing instance was already
running (init has pid 1 after all).

Thanks to estebistec!