Ability to use one log file per child process using format keys %i/%I
Starting from 3.2 there will be one log file per process, but that would be
useful even in this verison and is very hard to accomplish by configuration,
so this patch introduces a new format specifier for the logfile name.

Note that the numbers will stay within the process limit even if processes exit or
if autoscale/maxtasksperchild/time limits are used.  I.e. the number is the
*process index* not the process count or pid.

The new format specifiers are:

* ``%i`` -  Pool process index or 0 if MainProcess.

    With ``-n worker1@example.com -c2 -f %n-%i.log`` this will result in
    three logfiles:

        - ``worker1-0.log`` (main process)
        - ``worker1-1.log`` (pool process 1)
        - ``worker1-2.log`` (pool process 2)

* ``%I`` -  Pool process index with separator.

    With ``-n worker1@example.com -c2 -f %n%i.log` this will result in
    three logfiles:

        - ``worker1.log`` (main process)
        - ``worker1-1.log`` (pool process 1)
        - ``worker1-2.log`` (pool process 2)