Teach autodoc to document tasks if undoc-members is not set (#4588)
Celery tasks created with the @task decorator have the property
that `obj.__doc__` and `obj.__class__.__doc__` are equal, which
trips up the logic in `sphinx.ext.autodoc` that is supposed to
suppress repetition of class documentation in an instance of the
class. We have to override that behavior so that autodoc picks
up tasks whether or not autodoc's `undoc-members` flag is set.

Because autodoc really works now, we also have to correctly
check that the task was defined in the current module!

Fixes #4587.