Fix Sphinx support for shared_task decorated functions (#5135)
Currently, functions decorated with `@app.task` are picked up by
Sphinx's `automodule` directive, but those decorated with `@shared_task`
are not. This is related to the use of `PromiseProxy` in `app.task` vs.
`Proxy` in `shared_task` and how each of those handle `__module__`. This
change allows `celery.contrib.sphinx` to detect and render `shared_task`
functions as well.